reduce nested file structure, lint fix
