+++ src/sycl/kernels/gemm_sqrsum/device/gemm_sqrsum_types.hpp
@@ -9,13 +9,83 @@
 #include <sycl/sycl.hpp>
 
 #include "../../../Utils.h"
+#include "cutlass/cutlass.h"
 #include "cutlass/gemm/collective/collective_mma.hpp"
 #include "cutlass/numeric_types.h"
+#include "sycl/comm/common.h"
 #include "sycl/kernels/gemm_sqrsum/collective/xe_gemm_sqrsum_mainloop.hpp"
-#include "sycl/kernels/gemm_sqrsum/device/gemm_sqrsum_runner.hpp"
 #include "sycl/kernels/gemm_sqrsum/kernel/xe_gemm_sqrsum_kernel.hpp"
 
 using namespace cute;
+
+namespace cutlass::gemm_sqrsum::device {
+
+template <class Kernel_>
+class GemmSqrSum {
+ public:
+  using Kernel = Kernel_;
+  using Arguments = typename Kernel::Arguments;
+  using Params = typename Kernel::Params;
+
+ private:
+  Params params_;
+  bool initialized_ = false;
+
+ public:
+  GemmSqrSum() = default;
+
+  Params const& params() const {
+    return params_;
+  }
+
+  static cutlass::Status can_implement(Arguments const& args) {
+    return Kernel::can_implement(args) ? cutlass::Status::kSuccess : cutlass::Status::kErrorInvalidProblem;
+  }
+
+  static size_t get_workspace_size(Arguments const& args) {
+    return Kernel::get_workspace_size(args);
+  }
+
+  cutlass::Status initialize(
+      Arguments const& args, void* workspace = nullptr, sycl::queue& queue = c10::xpu::getCurrentXPUStream().queue()) {
+    params_ = Kernel::to_underlying_arguments(args, workspace);
+    initialized_ = true;
+    return cutlass::Status::kSuccess;
+  }
+
+  cutlass::Status update(Arguments const& args, void* workspace = nullptr) {
+    return initialize(args, workspace);
+  }
+
+  static cutlass::Status run(Params& params, sycl::queue& queue = c10::xpu::getCurrentXPUStream().queue()) {
+    launch<Kernel, 256>(params);
+    return cutlass::Status::kSuccess;
+  }
+
+  cutlass::Status
+  run(Arguments const& args, void* workspace = nullptr, sycl::queue& queue = c10::xpu::getCurrentXPUStream().queue()) {
+    cutlass::Status status = initialize(args, workspace, queue);
+    if (cutlass::Status::kSuccess == status) {
+      status = run(params_, queue);
+    }
+    return status;
+  }
+
+  cutlass::Status operator()(
+      Arguments const& args, void* workspace = nullptr, sycl::queue& queue = c10::xpu::getCurrentXPUStream().queue()) {
+    return run(args, workspace, queue);
+  }
+
+  cutlass::Status run(sycl::queue& queue = c10::xpu::getCurrentXPUStream().queue()) {
+    return run(params_, queue);
+  }
+
+  cutlass::Status operator()(sycl::queue& queue = c10::xpu::getCurrentXPUStream().queue()) {
+    return run(params_, queue);
+  }
+};
+
+}  // namespace cutlass::gemm_sqrsum::device
 
 struct GemmSqrSumXe {
   using Element = cutlass::tfloat32_t;
